+++ backend/sky-server/src/main/java/com/sky/service/DishService.java
@@ -43,5 +43,5 @@ public interface DishService {
     /**
      * 根据分类id查询菜品
      */
-    List<Dish> getListByCategoryId(Integer categoryId);
+    List<Dish> getListByCategoryId(Long categoryId);
 }
